+++ src/creatReact.js
@@ -4,14 +4,17 @@ import { render } from 'react-dom';
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import { createStore, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
+import { createBrowserHistory } from 'history';
 
 import { isFunction, isArray } from './utils';
 import creatReducers from './creatReducers';
 import creatStore from './creatStore';
+import creatSub from './creatSubscriptions';
 
 class creactReact {
 	constructor() {
 		this.router = '';
+		this.subs = ''; // subscriptions
 	}
 
 	// 获取路由
@@ -19,46 +22,76 @@ class creactReact {
 		this.router = router;
 	}
 
-	// 创建reducers
+	// 获取model，处理model
 	// @creatReducers
-	getReducers(params) {
+	getModel(params) {
+		let d = '';
 		if (isArray(params)) {
 			const reducersArr = [];
+			const subArr = [];
 			let buildReducers = '';
 			params.forEach(item => {
 				const {
 					namespace,
 					state,
 					reducers,
+					subscriptions,
 				} = item;
 				buildReducers = creatReducers(namespace, state, reducers);
 
+				// 获取所以reducer
 				reducersArr.push(buildReducers);
+
+				const subObj = {
+					namespace,
+					reducers,
+					subscriptions,
+				};
+
+				// 获取subscriptions
+				subArr.push(subObj);
 			});
 			// 创建store
-			const d = creatStore(reducersArr);
+			d = creatStore(reducersArr);
 
-			this.store = createStore(d);
+			// 存储subscriptions
+			this.subs = subArr;
+
 		} else {
 			const {
 				namespace,
 				state,
 				reducers,
+				subscriptions,
 			} = params;
 			const buildReducers = creatReducers(namespace, state, reducers);
 
 			// 创建store
-			const d = creatStore(buildReducers);
+			d = creatStore(buildReducers);
 
-			this.store = createStore(d);
+			const subObj = {
+				namespace,
+				reducers,
+				subscriptions
+			};
+			// 存储subscriptions
+			this.subs = subObj;
 		}
-		
+
+		this.store = createStore(d);
 	}
 
 	// 创建渲染
 	creat(selector) {
+
+		// 事件订阅
+		creatSub(this.subs, this.store);
+		
 		if (isFunction(this.router)) {
-			render(<Provider store={this.store}>{this.router()}</Provider>, selector);
+			render(
+				<Provider store={this.store}>
+					{this.router({ history: createBrowserHistory() })}
+				</Provider>, selector);
 		} else if (isArray(this.router)) {
 			render(
 				<Provider store={this.store}>