+++ src/creatReducers.js
@@ -1,8 +1,8 @@
 import { returnSelf } from './utils';
 
-const handleAction = (actionType, reducer = returnSelf) => (state, action) => {
+const handleAction = (actionType, reducer = returnSelf, namespace) => (state, action) => {
 	const { type } = action;
-	if (actionType === type) {
+	if (actionType === type || `${namespace}/${actionType}` === type) {
 		return reducer(state, action);
 	}
 	return state;
@@ -14,11 +14,10 @@ const reduceReducers = (...reducers) =>
 
 export default (namespace, initialState, handlers) => {
 	const reducers = Object.keys(handlers).map(type =>
-		handleAction(type, handlers[type])
+		handleAction(type, handlers[type], namespace)
 	);
     
 	const reducer = reduceReducers(...reducers);
 	return { [namespace]: (state = initialState, action) => reducer(state, action) };
 };
-