+++ HISTORY.md
@@ -1,3 +1,15 @@
+## 1.0.0-beta6
+- router 带入history
+- redux subscriptions 添加
+
+## 1.0.0
+- react & redux & react-router-dom
+
+## 1.0.0-beta5
+- 对外放出router
+- import { reactRouterDom } from 'mido-react';
+- const { BrowserRouter, Route, Link } = reactRouterDom;
+
 ## 1.0.0-beta4
 - 内置react, redux, reactRedux
 - import { react, redux, reactRedux } from 'mido-react';